Lots of refinements..     - teslaCharger.cpp: More accurate AC power calculations for charger setpoint     - throttle.cpp: Clear errors automatically     - i3LIM.cpp:         - Added enums for pilot and plug status for clarity         - Used different bit in CAN msg 0x3B4 for plug status (more reliable/stable)     - F30_Lever.cpp: Default lever position is now PARK when turning on car instead of neutral     - EvControlsT2C.cpp: type fix for clarity     - DilithiumMCU.cpp: type and scaling fix for BMS cell temperatures     - BMW_E90.cpp:         - Fixed interval issue in consumption calculations         - Fixed vehicle speed (was in mp/h, now km/h)

+++ include/DilithiumMCU.h
@@ -39,8 +39,8 @@ class DilithiumMCU: public BMS
       int timeoutCounterGFM = 0;
       float minCellV = 0;
       float maxCellV = 0;
-      float minTempC = 0;
-      float maxTempC = 0;
+      int8_t minTempC = 0;
+      int8_t maxTempC = 0;
       float battVoltage = 0; // udc2 is battery voltage, udc is bus voltage
       float deltaV = 0;
       float power = 0;
